+++ src/main/java/com/airesnor/wuxiacraft/items/ItemScroll.java
@@ -56,7 +56,7 @@ public class ItemScroll extends Item implements IHasModel {
 			tooltip.add(line);
 		}
 		if (technique.getBaseModifiers().armor != 0) {
-			float armor = technique.getBaseModifiers().armor;
+			double armor = technique.getBaseModifiers().armor;
 			int signals = (int) Math.abs(armor / 0.1f);
 			String color = armor < 0 ? TextFormatting.RED.toString() : signals < 3 ? TextFormatting.GREEN.toString() : TextFormatting.AQUA.toString();
 			String line = color + "Armor ";
@@ -66,7 +66,7 @@ public class ItemScroll extends Item implements IHasModel {
 			tooltip.add(line);
 		}
 		if (technique.getBaseModifiers().attackSpeed != 0) {
-			float attackSpeed = technique.getBaseModifiers().attackSpeed;
+			double attackSpeed = technique.getBaseModifiers().attackSpeed;
 			int signals = (int) Math.abs(attackSpeed / 0.1f);
 			String color = attackSpeed < 0 ? TextFormatting.RED.toString() : signals < 3 ? TextFormatting.GREEN.toString() : TextFormatting.AQUA.toString();
 			String line = color + "Attack Speed ";
@@ -76,7 +76,7 @@ public class ItemScroll extends Item implements IHasModel {
 			tooltip.add(line);
 		}
 		if (technique.getBaseModifiers().maxHealth != 0) {
-			float maxHealth = technique.getBaseModifiers().maxHealth;
+			double maxHealth = technique.getBaseModifiers().maxHealth;
 			int signals = (int) Math.abs(maxHealth / 0.1f);
 			String color = maxHealth < 0 ? TextFormatting.RED.toString() : signals < 3 ? TextFormatting.GREEN.toString() : TextFormatting.AQUA.toString();
 			String line = color + "Max Health ";
@@ -86,7 +86,7 @@ public class ItemScroll extends Item implements IHasModel {
 			tooltip.add(line);
 		}
 		if (technique.getBaseModifiers().movementSpeed != 0) {
-			float movementSpeed = technique.getBaseModifiers().movementSpeed;
+			double movementSpeed = technique.getBaseModifiers().movementSpeed;
 			int signals = (int) Math.abs(movementSpeed / 0.1f);
 			String color = movementSpeed < 0 ? TextFormatting.RED.toString() : signals < 3 ? TextFormatting.GREEN.toString() : TextFormatting.AQUA.toString();
 			String line = color + "Speed ";
@@ -96,7 +96,7 @@ public class ItemScroll extends Item implements IHasModel {
 			tooltip.add(line);
 		}
 		if (technique.getBaseModifiers().strength != 0) {
-			float strength = technique.getBaseModifiers().strength;
+			double strength = technique.getBaseModifiers().strength;
 			int signals = (int) Math.abs(strength / 0.1f);
 			String color = strength < 0 ? TextFormatting.RED.toString() : signals < 3 ? TextFormatting.GREEN.toString() : TextFormatting.AQUA.toString();
 			String line = color + "Strength ";
